re organized the code and moved to  expenses folder
