+++ src/components/Expenses/Expenses.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ExpenseItem from './ExpenseItem';
 import './Expenses.css';
-import Card from './Card';
+import Card from '../Ui/Card';
 
 const Expenses = (props) => {
   return (
